page division in bottom bar
Podzieleni kodu w ten sposób że każda ze stron którą możemy otworzyć na pana zupełnym znajduje się w oddzielnym pliku - przygotowanie do późniejszego oddzielnego rozwoju innych stron
